EPMEDU-1079  split DataStorePreferencesKey
